pyproject.toml
[tool.poetry]
name = "favorites-api"
version = "0.1.0"
description = "Luizalabs Technical challenge - Favorites-API is the back-end responsible for storing the favorite products of the Magalu customers."
authors = ["Lucas de Oliveira <lucasvitalate@gmail.com>"]

[tool.black]
line-length = 110

[tool.isort]
line_length = 110
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
known_local_folder = "apps,utils,favorites_api,tests"
sections = "FUTURE,STDLIB,THIRDPARTY,FIRSTPARTY,LOCALFOLDER"
default_section = "THIRDPARTY"

[tool.poetry.dependencies]
python = "^3.9"
Django = "^3.1.3"
environs = "^9.2.0"
djangorestframework = "^3.12.2"
dj-database-url = "^0.5.0"
psycopg2 = "^2.8.6"
django-redis = "^4.12.1"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
pre-commit = "^2.8.2"
black = "^20.8b1"
isort = "^5.6.4"
safety = "^1.9.0"
pytest-django = "^4.1.0"
factory-boy = "^3.1.0"
responses = "^0.12.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
